Add MIT license and package-lock.json file
This commit includes the addition of the MIT license to the project. The MIT License file gives all necessary permissions including copying, modifying, merging, publishing, distributing, sublicensing, and selling copies of Software. It also contains the necessary conditions and disclaimers. In addition, it covers the creation of the package-lock.json file, which ensures that the installed dependencies match across all environments.